requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
langchain==1.1.0
langchain-groq==1.0.1
langchain-core==1.1.0
pydantic==2.7.4
groq==0.36.0
python-dateutil==2.8.2
cryptography==41.0.7
google-api-python-client==2.187.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-auth==2.41.1
sounddevice==0.4.6
scipy==1.11.4
streamlit==1.51.0
gtts==2.5.0
requests==2.31.0
pydub==0.25.1
soundfile==0.13.1